requirements.txt
matplotlib==3.7.5
tqdm==4.66.2
numpy==1.24.3
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2